work with Object::Remote by removing require() call on Tie::StdHash in PP
Since Tie::StdHash is in Tie::Hash, and base.pm triggers a require call on
it, Object::Remote gets confused and tries to send Tie/StdHash.pm over the
wire, which of course does not exist. Changing the `use base` call to
`our @ISA` avoids that and makes BHEOS work fine with Object::Remote again.